refactor class XTB
