+++ .github/workflows/pylint.yml
@@ -23,4 +23,4 @@ jobs:
         if [[ -f requirements.txt ]]; then pip install -r requirements.txt; fi
     - name: Analysing the code with pylint
       run: |
-        pylint $(git ls-files 'xtb_api.py') --indent-string='\t'
+        pylint $(git ls-files '*.py') --indent-string='\t'
